+++ src/main/kotlin/no/ok/origo/dataplatform/commons/auth/ClientCredentialsProvider.kt
@@ -48,7 +48,7 @@ class ClientCredentialsProvider(
         val formData = listOf(
             "client_id" to clientId,
             "client_secret" to clientSecret,
-            "refresh_token" to _token.refreshToken,
+            "refresh_token" to _token.refreshToken!!,
             "grant_type" to "refresh_token"
         )
         return client.tokenRequest(parameters = formData)